% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignTune-package.R
\docType{package}
\name{gsDesignTune-package}
\alias{gsDesignTune}
\alias{gsDesignTune-package}
\title{gsDesignTune: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

}
\keyword{internal}
