% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending-spec.R
\name{spending_specs}
\alias{spending_specs}
\title{Spending function specifications}
\description{
\code{SpendingSpec} and \code{SpendingFamily} provide a dependency-aware and
user-friendly way to tune spending functions and their parameters.
}
