% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending-spec.R
\name{SpendingFamily}
\alias{SpendingFamily}
\title{SpendingFamily}
\value{
An R6 class generator. Use \verb{$new()} to create a \code{SpendingFamily} object.
}
\description{
SpendingFamily

SpendingFamily
}
\details{
An R6 class representing a set of spending function specifications. Each
family member is a \code{SpendingSpec}.
}
\examples{
fam <- SpendingFamily$new(
  SpendingSpec$new(gsDesign::sfHSD, par = tune_fixed(-4)),
  SpendingSpec$new(gsDesign::sfLDOF, par = tune_fixed(0))
)
fam$expand()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{members}}{List of \code{SpendingSpec} objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpendingFamily-new}{\code{SpendingFamily$new()}}
\item \href{#method-SpendingFamily-expand}{\code{SpendingFamily$expand()}}
\item \href{#method-SpendingFamily-clone}{\code{SpendingFamily$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpendingFamily-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpendingFamily-new}{}}}
\subsection{Method \code{new()}}{
Create a new spending family from one or more \code{SpendingSpec}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpendingFamily$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{SpendingSpec} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpendingFamily-expand"></a>}}
\if{latex}{\out{\hypertarget{method-SpendingFamily-expand}{}}}
\subsection{Method \code{expand()}}{
Expand all members to spending settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpendingFamily$expand()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpendingFamily-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpendingFamily-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpendingFamily$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
