% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-specs.R
\name{tune_choice}
\alias{tune_choice}
\title{Categorical choices}
\usage{
tune_choice(...)
}
\arguments{
\item{...}{Candidate values.}
}
\value{
A \code{gstune_spec} object.
}
\description{
\code{tune_choice()} defines a finite set of categorical choices. Each argument in
\code{...} is treated as one choice (including functions and other objects).
}
\examples{
tune_choice("A", "B")
}
