% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-specs.R
\name{tune_fixed}
\alias{tune_fixed}
\title{Fixed (non-tuned) value}
\usage{
tune_fixed(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A \code{gstune_spec} object.
}
\description{
Use \code{tune_fixed()} to explicitly mark a value as fixed. This is mainly useful
inside dependent specifications such as \code{tune_dep()}.
}
\examples{
tune_fixed(0.025)
}
