% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-specs.R
\name{tune_int}
\alias{tune_int}
\title{Integer sequence candidates}
\usage{
tune_int(from, to, by = 1)
}
\arguments{
\item{from, to}{Integer scalars.}

\item{by}{Integer scalar step size.}
}
\value{
A \code{gstune_spec} object.
}
\description{
Integer sequence candidates
}
\examples{
tune_int(2, 5)
}
