% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-specs.R
\name{tune_seq}
\alias{tune_seq}
\title{Numeric sequence candidates}
\usage{
tune_seq(from, to, length_out)
}
\arguments{
\item{from, to}{Numeric scalars.}

\item{length_out}{Integer scalar, the number of candidates.}
}
\value{
A \code{gstune_spec} object.
}
\description{
Numeric sequence candidates
}
\examples{
tune_seq(0.55, 0.75, length_out = 5)
}
