% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-specs.R
\name{tune_specs}
\alias{tune_specs}
\title{Tune specifications}
\description{
\code{gsDesignTune()} and \code{gsSurvTune()} treat most arguments as fixed values.
Wrap an argument in a \verb{tune_*()} specification to explore candidate values.
}
