% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignTune-package.R
\docType{package}
\name{gsDesignTune-package}
\alias{gsDesignTune-package}
\title{gsDesignTune: Dependency-Aware Scenario Exploration for Group Sequential Designs}
\description{
Provides systematic, dependency-aware exploration of group sequential designs created with 'gsDesign'. Supports reproducible grid and random search over user-defined candidate sets, parallel evaluation via the 'future' framework, standardized metric extraction, and auditable reporting for design-space evaluation and trade-off analysis. Methods for group sequential design are described in Anderson (2025) \doi{10.32614/CRAN.package.gsDesign}. The 'future' framework for parallel processing is described in Bengtsson (2021) \doi{10.32614/RJ-2021-048}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/gsDesignTune/}
  \item \url{https://github.com/nanxstats/gsDesignTune}
  \item Report bugs at \url{https://github.com/nanxstats/gsDesignTune/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

}
\keyword{internal}
