% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{gsSurvTune}
\alias{gsSurvTune}
\title{Create a tune job for \code{gsDesign::gsSurv()}}
\usage{
gsSurvTune(..., upper = NULL, lower = NULL)
}
\arguments{
\item{...}{Arguments to \code{\link[gsDesign:nSurv]{gsDesign::gsSurv()}}. Any argument can be replaced
by a \verb{tune_*()} specification.}

\item{upper, lower}{Optional spending specifications provided as
\code{SpendingSpec} or \code{SpendingFamily}. When supplied, these are translated to
the underlying \verb{(sfu, sfupar)} / \verb{(sfl, sflpar)} arguments.}
}
\value{
A \code{GSDTuneJob} R6 object.
}
\description{
Create a tune job for \code{gsDesign::gsSurv()}
}
